1. Histogram & Frequency Array
Design and Implement a program in C on Frequency Histogram, that builds a frequency array for data values 
in the range 1 to n and then prints their histogram.
The program should,
a. Read, Store and Print the data in an array.
b. Analyze the data in the array, one element at a time. Add 1 to the corresponding element in a frequency array 
based on the data value.
c. Print a histogram using asterisks for each occurrence of an element.